+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix a false positive for `RSpec/VariableName` when inside non-spec code. ([@ydah])
 - Fix a false positive for `RSpec/VariableDefinition` when inside non-spec code. ([@ydah])
 - Add new `RSpec/PendingBlockInsideExample` cop. ([@ydah])
+- Add `RSpec/RedundantAround` cop. ([@r7kamura])
 
 ## 2.18.1 (2023-01-19)